requirements.txt
transformers 
imagecorruptions
pytorch-ignite
neptune
sentencepiece
evaluate
jiwer
srsly
imgaug
pydantic
albumentations